backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary
Flask-Migrate 
Werkzeug

